Add chat_gpt module
